Moved pointer code to different script
- Cleared projectileShoot script by moving pointer code
