+++ Assets/scenes/level 3.unity
@@ -4365,30 +4365,11 @@ Prefab:
       propertyPath: m_RootOrder
       value: 2
       objectReference: {fileID: 0}
-    - target: {fileID: 114921812004129910, guid: 5e1a4dd8042d74d7396bb196ddfbbde0,
-        type: 2}
-      propertyPath: leftCatapultLine
-      value: 
-      objectReference: {fileID: 1567680018}
-    - target: {fileID: 114921812004129910, guid: 5e1a4dd8042d74d7396bb196ddfbbde0,
-        type: 2}
-      propertyPath: rightCatapultLine
-      value: 
-      objectReference: {fileID: 1669298760}
     - target: {fileID: 231365880645621072, guid: 5e1a4dd8042d74d7396bb196ddfbbde0,
         type: 2}
       propertyPath: m_ConnectedRigidBody
       value: 
       objectReference: {fileID: 1714009865}
-    - target: {fileID: 50435492825116804, guid: 5e1a4dd8042d74d7396bb196ddfbbde0,
-        type: 2}
-      propertyPath: m_BodyType
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1194249144360712, guid: 5e1a4dd8042d74d7396bb196ddfbbde0, type: 2}
-      propertyPath: m_IsActive
-      value: 1
-      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 5e1a4dd8042d74d7396bb196ddfbbde0, type: 2}
   m_IsPrefabParent: 0
